Flatten Math and Geometry (#121)
