+++ src/Bearded.Utilities.csproj
@@ -130,21 +130,21 @@
     <Compile Include="Linq\Extensions.cs" />
     <Compile Include="Input\KeyboardAction.cs" />
     <Compile Include="Input\LambdaAction.cs" />
-    <Compile Include="Math\Angle.cs" />
-    <Compile Include="Math\Direction2.cs" />
-    <Compile Include="Math\Arc.cs" />
+    <Compile Include="Geometry\Angle.cs" />
+    <Compile Include="Geometry\Direction2.cs" />
+    <Compile Include="Geometry\arcs\Arc.cs" />
     <Compile Include="Core\MathExtensions.cs" />
-    <Compile Include="Math\GeometryExtensions.cs" />
-    <Compile Include="Math\Interval.cs" />
+    <Compile Include="Geometry\GeometryExtensions.cs" />
+    <Compile Include="Geometry\Interval.cs" />
     <Compile Include="Core\Mathf.cs" />
-    <Compile Include="Math\Geometry\arcs\Arc2.cs" />
-    <Compile Include="Math\Geometry\arcs\Arc3.cs" />
-    <Compile Include="Math\Geometry\arcs\Bezier2nd2.cs" />
-    <Compile Include="Math\Geometry\arcs\Bezier2nd3.cs" />
-    <Compile Include="Math\Geometry\arcs\Bezier3rd2.cs" />
-    <Compile Include="Math\Geometry\arcs\Bezier3rd3.cs" />
-    <Compile Include="Math\Geometry\PolarPosition.cs" />
-    <Compile Include="Math\Geometry\Rectangle.cs" />
+    <Compile Include="Geometry\arcs\Arc2.cs" />
+    <Compile Include="Geometry\arcs\Arc3.cs" />
+    <Compile Include="Geometry\arcs\Bezier2nd2.cs" />
+    <Compile Include="Geometry\arcs\Bezier2nd3.cs" />
+    <Compile Include="Geometry\arcs\Bezier3rd2.cs" />
+    <Compile Include="Geometry\arcs\Bezier3rd3.cs" />
+    <Compile Include="Geometry\PolarPosition.cs" />
+    <Compile Include="Geometry\Rectangle.cs" />
     <Compile Include="Core\Interpolate.cs" />
     <Compile Include="SpaceTime\angular\AngularAcceleration.cs" />
     <Compile Include="SpaceTime\angular\AngularVelocity.cs" />
