+++ src/Geometry/Direction2.cs
@@ -2,7 +2,7 @@ using System;
 using System.Globalization;
 using OpenTK;
 
-namespace Bearded.Utilities.Math
+namespace Bearded.Utilities.Geometry
 {
     /// <summary>
     /// A typesafe representation of a direction in two dimensional space.
