+++ src/Geometry/Interval.cs
@@ -1,4 +1,4 @@
-﻿namespace Bearded.Utilities.Math
+﻿namespace Bearded.Utilities.Geometry
 {
     /// <summary>
     /// A collection of interval helper methods.
