+++ src/Geometry/arcs/Arc3.cs
@@ -1,6 +1,6 @@
 ﻿using OpenTK;
 
-namespace Bearded.Utilities.Math.Geometry
+namespace Bearded.Utilities.Geometry
 {
     /// <summary>
     /// Represents an arc in three-dimensional Euclidean space.
