+++ src/Geometry/arcs/Bezier2nd3.cs
@@ -1,21 +1,21 @@
 ﻿using OpenTK;
 
-namespace Bearded.Utilities.Math.Geometry
+namespace Bearded.Utilities.Geometry
 {
     /// <summary>
-    /// Represents a quadratic Bezier curve in two-dimensional space.
+    /// Represents a quadratic Bezier curve in three-dimensional space.
     /// </summary>
     // ReSharper disable once InconsistentNaming
-    public sealed class Bezier2nd2 : Arc2
+    public sealed class Bezier2nd3 : Arc3
     {
-        private readonly Vector2 p0;
-        private readonly Vector2 p1;
-        private readonly Vector2 p2;
+        private readonly Vector3 p0;
+        private readonly Vector3 p1;
+        private readonly Vector3 p2;
 
         /// <summary>
         /// Initializes the Bezier curve with 100 segments.
         /// </summary>
-        public Bezier2nd2(Vector2 p0, Vector2 p1, Vector2 p2)
+        public Bezier2nd3(Vector3 p0, Vector3 p1, Vector3 p2)
             : this(p0, p1, p2, 100)
         {
         }
@@ -23,7 +23,7 @@ namespace Bearded.Utilities.Math.Geometry
         /// <summary>
         /// Initializes the Bezier curve.
         /// </summary>
-        public Bezier2nd2(Vector2 p0, Vector2 p1, Vector2 p2, int segments)
+        public Bezier2nd3(Vector3 p0, Vector3 p1, Vector3 p2, int segments)
             : base(segments)
         {
             this.p0 = p0;
@@ -36,7 +36,7 @@ namespace Bearded.Utilities.Math.Geometry
         /// </summary>
         /// <param name="t">The arc parameter t.</param>
         /// <returns>The Euclidean coordinates of the point on the curve at parameter t.</returns>
-        protected override Vector2 getPointAt(float t)
+        protected override Vector3 getPointAt(float t)
         {
             return Interpolate.Bezier(p0, p1, p2, t);
         }
