+++ src/Geometry/arcs/Bezier3rd2.cs
@@ -1,34 +1,36 @@
 ﻿using OpenTK;
 
-namespace Bearded.Utilities.Math.Geometry
+namespace Bearded.Utilities.Geometry
 {
     /// <summary>
-    /// Represents a quadratic Bezier curve in three-dimensional space.
+    /// Represents a cubic Bezier curve in two-dimensional space.
     /// </summary>
     // ReSharper disable once InconsistentNaming
-    public sealed class Bezier2nd3 : Arc3
+    public sealed class Bezier3rd2 : Arc2
     {
-        private readonly Vector3 p0;
-        private readonly Vector3 p1;
-        private readonly Vector3 p2;
+        private readonly Vector2 p0;
+        private readonly Vector2 p1;
+        private readonly Vector2 p2;
+        private readonly Vector2 p3;
 
         /// <summary>
         /// Initializes the Bezier curve with 100 segments.
         /// </summary>
-        public Bezier2nd3(Vector3 p0, Vector3 p1, Vector3 p2)
-            : this(p0, p1, p2, 100)
+        public Bezier3rd2(Vector2 p0, Vector2 p1, Vector2 p2, Vector2 p3)
+            : this(p0, p1, p2, p3, 100)
         {
         }
 
         /// <summary>
         /// Initializes the Bezier curve.
         /// </summary>
-        public Bezier2nd3(Vector3 p0, Vector3 p1, Vector3 p2, int segments)
+        public Bezier3rd2(Vector2 p0, Vector2 p1, Vector2 p2, Vector2 p3, int segments)
             : base(segments)
         {
             this.p0 = p0;
             this.p1 = p1;
             this.p2 = p2;
+            this.p3 = p3;
         }
 
         /// <summary>
@@ -36,9 +38,9 @@ namespace Bearded.Utilities.Math.Geometry
         /// </summary>
         /// <param name="t">The arc parameter t.</param>
         /// <returns>The Euclidean coordinates of the point on the curve at parameter t.</returns>
-        protected override Vector3 getPointAt(float t)
+        protected override Vector2 getPointAt(float t)
         {
-            return Interpolate.Bezier(p0, p1, p2, t);
+            return Interpolate.Bezier(p0, p1, p2, p3, t);
         }
     }
 }
