+++ src/Geometry/arcs/Bezier3rd3.cs
@@ -1,22 +1,22 @@
 ﻿using OpenTK;
 
-namespace Bearded.Utilities.Math.Geometry
+namespace Bearded.Utilities.Geometry
 {
     /// <summary>
-    /// Represents a cubic Bezier curve in two-dimensional space.
+    /// Represents a cubic Bezier curve in three-dimensional space.
     /// </summary>
     // ReSharper disable once InconsistentNaming
-    public sealed class Bezier3rd2 : Arc2
+    public sealed class Bezier3rd3 : Arc3
     {
-        private readonly Vector2 p0;
-        private readonly Vector2 p1;
-        private readonly Vector2 p2;
-        private readonly Vector2 p3;
+        private readonly Vector3 p0;
+        private readonly Vector3 p1;
+        private readonly Vector3 p2;
+        private readonly Vector3 p3;
 
         /// <summary>
         /// Initializes the Bezier curve with 100 segments.
         /// </summary>
-        public Bezier3rd2(Vector2 p0, Vector2 p1, Vector2 p2, Vector2 p3)
+        public Bezier3rd3(Vector3 p0, Vector3 p1, Vector3 p2, Vector3 p3)
             : this(p0, p1, p2, p3, 100)
         {
         }
@@ -24,7 +24,7 @@ namespace Bearded.Utilities.Math.Geometry
         /// <summary>
         /// Initializes the Bezier curve.
         /// </summary>
-        public Bezier3rd2(Vector2 p0, Vector2 p1, Vector2 p2, Vector2 p3, int segments)
+        public Bezier3rd3(Vector3 p0, Vector3 p1, Vector3 p2, Vector3 p3, int segments)
             : base(segments)
         {
             this.p0 = p0;
@@ -38,7 +38,7 @@ namespace Bearded.Utilities.Math.Geometry
         /// </summary>
         /// <param name="t">The arc parameter t.</param>
         /// <returns>The Euclidean coordinates of the point on the curve at parameter t.</returns>
-        protected override Vector2 getPointAt(float t)
+        protected override Vector3 getPointAt(float t)
         {
             return Interpolate.Bezier(p0, p1, p2, p3, t);
         }
